Damage_Number
Damage_Number를 호출하려면

Enemy가 받은 데미지,(Player가 가한 데미지) 위치정보(Transform.position)으로 호출하시면 됩니다.*테스트 할때는 Enemy의 TakeDamage에서 amount와 transform.position을 활용했습니다. *

(UiManager의 EnemyDamageTextUi() 함수)
(ReadMe에도 적었습니다.)

+++ Assets/01.Scripts/UI/DamageUi.cs
@@ -14,7 +14,7 @@ public class DamageUi : MonoBehaviour
 
     private float timer;
 
-    public void Show(int damage, Vector3 position)
+    public void Show(float damage, Vector3 position)
     {
         transform.position = position;
